Working Portal Implementation
I also ran rustfmt

+++ src/js.rs
@@ -5,7 +5,11 @@ use node_bindgen::derive::node_bindgen;
 use crate::structs::{KeybindId, KeybindTrigger};
 
 #[node_bindgen]
-async fn start_keybinds<F: Fn(KeybindId)>(window_id: Option<usize>, display_id: Option<usize>, callback: F) {
+async fn start_keybinds<F: Fn(KeybindId)>(
+    window_id: Option<u64>,
+    display_id: Option<u64>,
+    callback: F,
+) {
     let (tx, rx) = channel::<KeybindTrigger>();
     thread::spawn(|| {
         crate::start_keybinds(window_id, display_id, tx);
